separate icmt from icmt-sdl2
