+++ icmt-sdl2/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "icmt"
+name = "icmt-sdl2"
 version = "0.1.0"
 authors = ["Matthew A Hammer <pubmah@nym.hush.com>"]
 edition = "2018"
@@ -10,9 +10,13 @@ repository = "https://github.com/matthewhammer/ic-mini-terminal"
 license = "Apache-2.0"
 readme = "README.md"
 
+[[bin]]
+name = "ic-mt"
+path = "src/bin/ic-mt.rs"
+
 [lib]
 crate-type = ["lib"]
-name = "icmt"
+name = "icmt_sdl2"
 path = "src/lib/mod.rs"
 test = true            # Is tested by default.
 doctest = true         # Documentation examples are tested by default.
@@ -34,16 +38,22 @@ serde = { version = "1.0", features = ["derive"] }
 serde_bytes = "0.11"
 serde_cbor = "0.9"
 serde_json = "1.0"
+sdl2 = "0.34.3"
 tokio = { version = "1", features = ["full"] }
 garcon = "0.2.3"
 hex = "0.4.2"
 num-bigint = "0.2.6"
 num-traits = "0.2.6"
 futures = "0.3.5"
+ring = "0.16.15"
+engiffen = "0.8.1"
+ic-agent = "0.5.0"
+ic-types = "0.1.3"
 #candid = "0.6.11"
 ron = "*"
 js-sys = "0.3"
 #wasm-bindgen = "0.2.58"
+icmt = { path = "../icmt" }
 
 [dependencies.candid]
 git = "https://github.com/dfinity/candid"
