+++ icmt-sdl2/src/lib/error.rs
